requirements.txt
protobuf==4.21.6
python-dotenv==0.21.0
firebase-admin==6.2.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
grpcio-status==1.59.2
google-auth-oauthlib==1.1.0
importlib-metadata==6.8.0
langchain==0.0.329
openai==0.28.1
openapi-schema-pydantic==1.2.4
openai-whisper
openpyxl==3.1.2
pandas-stubs==2.1.1.230928
pandas==2.1.2
pip>=23.3
plotly==5.18.0
pyOpenSSL==23.3.0
PySocks==1.7.1
pytube==15.0.0
pyu2f==0.1.5
scikit-learn==1.3.2
textblob==0.15.3
unicodedata2==15.1.0
chromadb==0.4.15
unstructured==0.10.28